Move `mem` into `js`.
