+++ src/js/binary.rs
@@ -4,8 +4,9 @@ use std::marker::PhantomData;
 use std::mem;
 use std::os::raw::c_void;
 use std::slice;
-use vm::{Context, JsResult, Managed};
+use vm::{Context, JsResult};
 use js::{Value, Object, Borrow, BorrowMut, Ref, RefMut, LoanError, build};
+use js::mem::Managed;
 use js::internal::ValueInternal;
 use vm::VmGuard;
 use vm::internal::Pointer;
