+++ src/prelude.rs
@@ -1,4 +1,4 @@
 //! The Neon "prelude," a re-exported collection of the most commonly-used Neon APIs.
 
-pub use js::{JsBuffer, JsArrayBuffer, BinaryData, Class, JsError, ErrorKind, Value, JsValue, JsUndefined, JsNull, JsBoolean, JsString, ToJsString, JsNumber, JsObject, Object, JsArray, JsFunction, Borrow, BorrowMut};
-pub use vm::{Handle, DowncastError, DowncastResult, VmResult, JsResult, JsResultExt, CallKind, Context, ModuleContext, ExecuteContext, ComputeContext, CallContext, FunctionContext, MethodContext};
+pub use js::{Handle, JsBuffer, JsArrayBuffer, BinaryData, Class, JsError, ErrorKind, Value, JsValue, JsUndefined, JsNull, JsBoolean, JsString, ToJsString, JsNumber, JsObject, Object, JsArray, JsFunction, Borrow, BorrowMut};
+pub use vm::{VmResult, JsResult, JsResultExt, CallKind, Context, ModuleContext, ExecuteContext, ComputeContext, CallContext, FunctionContext, MethodContext};
